+++ globals.py
@@ -1,0 +1,5 @@
+import numpy as np
+
+globals = {
+    "floatX" : np.float64
+}
